+++ custom_assistance_support/models/follower_group.py
@@ -1,8 +1,9 @@
-from odoo import models, fields, api
+from odoo import fields, models
+
 
 class FollowerGroup(models.Model):
-    _name = 'mail.follower.group'
-    _description = 'Follower Group'
+    _name = "mail.follower.group"
+    _description = "Follower Group"
 
     name = fields.Char(required=True)
-    partner_ids = fields.Many2many('res.partner', string='Contacts')
+    partner_ids = fields.Many2many("res.partner", string="Contacts")